Fix rounding error for length function of integral type vector.
`std::sqrt` with parameter of integral type will return double
nevertheless, which on conversion back to integral type will truncate
rather than round the result, as added in C++11. Old way to solve it was
to add 0.5f to the result, so the truncation yelds correct result, but
C++11 also added `std::round` for that reason.